Remove circular reference in services
